+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,18 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Pavel Barto on 12.04.2024.
+//
+
+import Foundation
+
+private struct QuizQuestion {
+    // строка с названием фильма,
+    // совпадает с названием картинки афиши фильма в Assets
+    let image: String
+    // строка с вопросом о рейтинге фильма
+    let text: String
+    // булевое значение (true, false), правильный ответ на вопрос
+    let correctAnswer: Bool
+}
