+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Pavel Barto on 12.04.2024.
+//
+
+import Foundation
+
+// для состояния "Результат квиза"
+private struct QuizResultsViewModel {
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
+}
