+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Pavel Barto on 12.04.2024.
+//
+
+import UIKit
+// вью модель для состояния "Вопрос показан"
+private struct QuizStepViewModel {
+    // картинка с афишей фильма с типом UIImage
+    let image: UIImage
+    // вопрос о рейтинге квиза
+    let question: String
+    // строка с порядковым номером этого вопроса (ex. "1/10")
+    let questionNumber: String
+}
